Put all bundled feature logic in one file
